07_Deep_Agents/ai_life_coach/pyproject.toml
[project]
name = "ai-life-coach"
version = "1.0.0"
description = "AI Life Coach - A Deep Agent system for comprehensive life guidance across career, relationships, finance, and wellness domains"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [
    {name = "AI Life Coach Project Team"}
]
keywords = ["ai", "agent", "life-coach", "deep-agents", "langchain", "wellness"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: End Users/Desktop",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    # Deep Agents framework
    "deepagents>=0.3.0",
    
    # LangChain ecosystem
    "langchain>=0.3.0",
    "langgraph>=1.0.0",
    
    # Model providers
    "langchain-openai>=0.2.0",
    "langchain-anthropic>=0.3.0",
    
    # Utilities
    "python-dotenv>=1.0.0",
    "pyyaml>=6.0",
    
    # Async support
    "nest-asyncio>=1.5.0",
]

[project.optional-dependencies]
# CLI for interactive agent sessions
cli = [
    "deepagents-cli>=0.0.15",
]

# Development tools
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
    "mypy>=1.5.0",
]

# Jupyter notebook support
notebook = [
    "jupyter>=1.0.0",
    "ipykernel>=6.0.0",
]

[project.scripts]
ai-life-coach = "src.main:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.black]
line-length = 100
target-version = ['py310', 'py311', 'py312']

[tool.ruff]
line-length = 100
select = ["E", "F", "I", "N", "W"]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true